Admin/app/requirements.txt
flask==2.0.1
python-dotenv==0.20.0